+++ src/webview/constants/Defaults.ts
@@ -1,8 +1,13 @@
+import { Gradients } from "./Gradients";
+
 export const Defaults = {
   fontSize: 16,
   innerWidth: 100,
   innerPadding: 2,
   innerBorder: 15,
   width: 1200,
-  height: 675
+  height: 675,
+  shadow: 25,
+  outerBackground: Gradients[3],
+  showTitleBar: false,
 }